chore: remove old table
remove old table component
