+++ app/components/Table.tsx
@@ -1,37 +1,25 @@
-import React, { CSSProperties, useState } from 'react';
-import styled from 'styled-components';
-import noop from 'lodash.noop';
-import Input from '@button-inc/bcgov-theme/Input';
-import Button from '@button-inc/bcgov-theme/Button';
-import Dropdown from '@button-inc/bcgov-theme/Dropdown';
+import { Table as StyledTable, SearchBar } from '@bcgov-sso/common-react-components';
+import React, { ReactElement, useEffect, useState } from 'react';
+import { useTable, usePagination, useFilters, useGlobalFilter, Column, useSortBy, Row, Cell } from 'react-table';
 import Grid from '@button-inc/bcgov-theme/Grid';
 import Pagination from 'react-bootstrap/Pagination';
-import { Table as StyledTable, SearchBar } from '@bcgov-sso/common-react-components';
-import SectionHeader from 'components/SectionHeader';
-import InfoOverlay from 'components/InfoOverlay';
-import { MultiSelect } from 'react-multi-select-component';
-import { TextBlock } from 'react-placeholder/lib/placeholders';
-import ReactPlaceholder from 'react-placeholder';
-import 'react-placeholder/lib/reactPlaceholder.css';
-import { TABLE_ROW_HEIGHT, TABLE_ROW_SPACING } from 'styles/theme';
+import styled from 'styled-components';
+import Select from 'react-select';
+import SectionHeader from './SectionHeader';
+import { faSortDown, faSortUp } from '@fortawesome/free-solid-svg-icons';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { Option } from 'interfaces/Request';
+import noop from 'lodash.noop';
+import InfoOverlay from './InfoOverlay';
+import ReactPlaceholder from 'react-placeholder/lib';
+import { TextBlock } from 'react-placeholder/lib/placeholders';
+import { TABLE_ROW_HEIGHT_MINI, TABLE_ROW_SPACING } from 'styles/theme';
+import InfiniteScroll from 'react-infinite-scroller';
 
-const StyledInput = styled(Input)`
-  input {
-    width: 100%;
-  }
-`;
+const InfScroll = InfiniteScroll as unknown as (a: any) => JSX.Element;
 
-const StyledMultiSelect = styled(MultiSelect)`
-  font-size: 0.9rem;
-
-  .dropdown-container {
-    border: 1.8px solid black !important;
-
-    .dropdown-heading {
-      height: 32px;
-    }
-  }
+const PaginationIcon = styled(FontAwesomeIcon)`
+  color: '#000';
 `;
 
 const StyledPagination = styled(Pagination)`
@@ -54,6 +42,11 @@ const PageInfo = styled.li`
   line-height: 40px;
 `;
 
+const StyledSelect = styled(Select)`
+  width: 150px;
+  display: inline-block;
+`;
+
 const FiltersContainer = styled.div<{ itemsLength: number }>`
   display: grid;
   grid-template-columns:
@@ -64,14 +57,22 @@ const FiltersContainer = styled.div<{ itemsLength: number }>`
   }
 `;
 
-export interface TableHeader {
-  label: string;
-  style?: CSSProperties;
-}
-
-export interface TableFilterItem {
-  value: string | number;
-  text: string;
+function SelectColumnFilter({ setFilter, options, setValue, gotoPage }: any) {
+  return (
+    <div data-testid="multi-select-col-filter">
+      <Select
+        className="basic-multi-select"
+        classNamePrefix="select"
+        onChange={(val) => {
+          setFilter('status', val);
+          setValue(val);
+          gotoPage(0);
+        }}
+        options={options}
+        isMulti
+      ></Select>
+    </div>
+  );
 }
 
 export interface TableFilter {
@@ -83,19 +84,66 @@ export interface TableFilter {
   label?: string;
 }
 
+export interface PaginationItemsDetailProps {
+  pageOptions: number[];
+  canPreviousPage: boolean;
+  canNextPage: boolean;
+  previousPage: () => void;
+  nextPage: () => void;
+  pageIndex: number;
+}
+
+const PaginationItemsDetail = ({
+  canPreviousPage,
+  previousPage,
+  canNextPage,
+  nextPage,
+  pageOptions = [],
+  pageIndex,
+}: PaginationItemsDetailProps) => {
+  return (
+    <>
+      <Pagination.Item
+        key="prev"
+        disabled={!canPreviousPage}
+        onClick={() => {
+          previousPage();
+        }}
+      >
+        Previous
+      </Pagination.Item>
+      <Pagination.Item
+        key="next"
+        disabled={!canNextPage}
+        onClick={() => {
+          nextPage();
+        }}
+      >
+        Next
+      </Pagination.Item>
+      <PageInfo>{`${pageIndex + 1} of ${pageOptions.length}`}</PageInfo>
+    </>
+  );
+};
+
+const getColumnSortedIcon = (isSorted: boolean, isSortedDesc: boolean | undefined) => {
+  if (isSorted) {
+    return <PaginationIcon icon={isSortedDesc ? faSortDown : faSortUp} size="sm" />;
+  }
+  return null;
+};
+
 interface Props {
   variant?: string;
-  headers: TableHeader[];
-  children: React.ReactNode;
-  pagination?: boolean;
-  pageLimits?: TableFilterItem[];
+  headers: Column[];
+  data: any;
+  activateRow?: (row: Row) => void;
+  loadMoreItem?: any;
+  hasMoreItem?: boolean;
   searchKey?: string;
   searchPlaceholder?: string;
   searchTooltip?: string;
-  page?: number;
-  limit?: number;
-  rowCount?: number;
-  filters: TableFilter[];
+  colfilters: TableFilter[];
   searchLocation?: 'left' | 'right';
   totalColSpan?: number;
   searchColSpan?: number;
@@ -104,110 +152,62 @@ interface Props {
   showContent?: boolean;
   headerAlign?: string;
   headerGutter?: number[];
-  onSearch?: (val: string) => void;
-  onEnter?: (val: string) => void;
-  onFilter?: (val: any) => void;
-  onFilter2?: (val: any) => void;
-  onLimit?: (val: number) => void;
-  onPage?: (val: number) => void;
-  onPrev?: (val: number) => void;
-  onNext?: (val: number) => void;
   loading?: boolean;
+  loader?: any;
+  onPage?: (val: number) => void;
+  rowCount?: number;
+  onEnter?: (val: string) => void;
+  onSearch?: (val: string) => void;
+  pageLimits?: number[];
+  onLimit?: (val: number) => void;
+  limit?: number;
+  noDataFoundElement?: ReactElement;
+  pagination?: boolean;
+  rowSelectorKey?: string;
+  readOnly?: boolean;
+  /**
+   * Pass in this prop if you want to control the selected row from the parent component. It should equal the rowSelectorKey's value in the active row. Leave undefined to control selected state internally. */
+  activeSelector?: string | number;
 }
-
-const overrideStrings = {
-  allItemsAreSelected: 'All',
-  selectSomeItems: '',
-};
-
-const generateOptions = (items: TableFilterItem[]) => (
-  <>
-    {items.map((item) => (
-      <option key={item.value} value={item.value}>
-        {item.text}
-      </option>
-    ))}
-  </>
-);
-
-const PaginationItemsDetail = ({ rowCount, limit, page, onPage, onPrev, onNext }: any) => {
-  const pageCount = parseInt(String((rowCount - 1) / limit + 1));
-
-  const items = [
-    <Pagination.Item key="prev" disabled={page === 1} onClick={() => onPrev()}>
-      Previous
-    </Pagination.Item>,
-  ];
-
-  for (let number = 1; number <= pageCount; number++) {
-    items.push(
-      <Pagination.Item key={number} active={number === page} onClick={() => onPage(number)}>
-        {number}
-      </Pagination.Item>,
-    );
-
-    if (pageCount === number) {
-      items.push(
-        <Pagination.Item key="next" disabled={page === number} onClick={() => onNext()}>
-          Next
-        </Pagination.Item>,
-      );
-    }
-  }
-
-  return <>{items}</>;
-};
-
-const PaginationItems = ({ rowCount, limit, page, onPrev, onNext }: any) => {
-  const pageCount = parseInt(String((rowCount - 1) / limit + 1));
-  const startNum = parseInt(String((page - 1) * limit + 1));
-  let endNum = startNum + limit - 1;
-  if (endNum > rowCount) endNum = rowCount;
-
-  return (
-    <>
-      <Pagination.Item key="prev" disabled={page === 1} onClick={() => onPrev(page - 1)}>
-        Previous
-      </Pagination.Item>
-      <Pagination.Item key="next" disabled={page === pageCount} onClick={() => onNext(page + 1)}>
-        Next
-      </Pagination.Item>
-      <PageInfo>{`${startNum}-${endNum} of ${rowCount}`}</PageInfo>
-    </>
-  );
-};
 
 function Table({
   variant = 'medium',
   headers,
-  children,
-  pagination = false,
-  onSearch = noop,
-  onEnter = noop,
-  filters = [] as TableFilter[],
+  data,
+  activateRow = noop,
+  loadMoreItem = noop,
+  hasMoreItem = false,
   searchLocation = 'left',
-  totalColSpan = 14,
-  searchColSpan = 4,
-  filterColSpan = 10,
-  showFilters = true,
-  showContent = true,
-  headerAlign = 'center',
-  headerGutter = [],
-  onLimit = noop,
-  onPage,
-  onPrev = noop,
-  onNext = noop,
-  pageLimits = [],
-  searchKey = '',
+  colfilters = [],
+  showFilters = false,
   searchPlaceholder = 'Search...',
   searchTooltip = '',
-  page = 1,
-  limit = 10,
+  totalColSpan = 14,
+  headerAlign = 'center',
+  headerGutter = [],
+  filterColSpan = 10,
+  searchColSpan = 4,
+  onPage = noop,
+  onEnter = noop,
   rowCount = 10,
+  searchKey = '',
+  onSearch = noop,
   loading,
+  loader = <></>,
+  onLimit = noop,
+  limit = 10,
+  noDataFoundElement = <p>No Data Found.</p>,
+  pagination = false,
+  pageLimits = [5, 10, 15, 30, 50, 100],
+  rowSelectorKey = 'id',
+  readOnly = false,
+  activeSelector,
 }: Props) {
+  const numOfPages = Math.ceil(rowCount / limit);
+  const columns: Column[] = React.useMemo(() => headers, [headers]);
+  const rowsData = React.useMemo(() => data, [data]);
+  const [selectedRow, setSelectedRow] = useState<any>();
   const [_searchKey, setSearchKey] = useState(searchKey);
-
   const handleSearchKeyChange = (event: React.ChangeEvent<HTMLInputElement>) => {
     setSearchKey(event.target.value);
   };
@@ -215,16 +215,45 @@ function Table({
   const handleKeyUp = (event: React.KeyboardEvent<HTMLInputElement>) => {
     if (event.key === 'Enter') {
       onEnter(_searchKey);
+      gotoPage(0);
     }
-  };
-
-  const handlePageLimitChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    onLimit(Number(event.target.value));
   };
 
   const handleSearchSubmit = () => {
     onSearch(_searchKey);
+    gotoPage(0);
+    setSelectedRow(undefined);
   };
+
+  const {
+    getTableProps,
+    getTableBodyProps,
+    headerGroups,
+    prepareRow,
+    page,
+    canPreviousPage,
+    canNextPage,
+    pageOptions,
+    nextPage,
+    previousPage,
+    state: { pageIndex },
+    setFilter,
+    gotoPage,
+  } = useTable(
+    {
+      columns,
+      data: rowsData,
+      initialState: { pageIndex: 0 },
+      manualPagination: true,
+      pageCount: numOfPages === 0 ? 1 : numOfPages,
+      autoResetPage: false,
+      autoResetSortBy: false,
+    },
+    useFilters,
+    useGlobalFilter,
+    useSortBy,
+    usePagination,
+  );
 
   let rowSpaces = rowCount || limit;
   if (rowCount > limit) rowSpaces = limit;
@@ -233,7 +262,7 @@ function Table({
     <td colSpan={100}>
       <div
         style={{
-          height: `${rowSpaces * (TABLE_ROW_HEIGHT + TABLE_ROW_SPACING) - TABLE_ROW_SPACING}px`,
+          height: `${rowSpaces * (TABLE_ROW_HEIGHT_MINI + TABLE_ROW_SPACING) - TABLE_ROW_SPACING}px`,
           paddingTop: '10px',
         }}
       >
@@ -241,6 +270,24 @@ function Table({
       </div>
     </td>
   );
+
+  const numOfItemsPerPage = () => {
+    const options = pageLimits.map((val) => {
+      return { value: val, label: `${val} per page` };
+    });
+
+    return options;
+  };
+
+  const handlePageLimitChange = (val: number) => {
+    gotoPage(0);
+    onLimit(val);
+  };
+
+  const updateSelectedRow = (row: any) => {
+    activateRow(row);
+    setSelectedRow(row);
+  };
 
   const searchCol = (
     <Grid.Col span={searchColSpan}>
@@ -258,11 +305,22 @@ function Table({
             />
           </Grid.Col>
           <Grid.Col span={4}>
-            <InfoOverlay content={searchTooltip}>
-              <Button type="button" size="small" onClick={handleSearchSubmit}>
+            {searchTooltip ? (
+              <InfoOverlay content={searchTooltip || 'some text'}>
+                <button
+                  className="primary"
+                  type="button"
+                  onClick={handleSearchSubmit}
+                  style={{ padding: '.44rem 1.5rem' }}
+                >
+                  Search
+                </button>
+              </InfoOverlay>
+            ) : (
+              <button className="primary short" type="button" onClick={handleSearchSubmit}>
                 Search
-              </Button>
-            </InfoOverlay>
+              </button>
+            )}
           </Grid.Col>
         </Grid.Row>
       </Grid>
@@ -271,41 +329,34 @@ function Table({
 
   const filterCol = (
     <Grid.Col span={filterColSpan} style={{ textAlign: 'right' }}>
-      <FiltersContainer itemsLength={filters.length}>
-        {filters.map((filter, index) => (
-          <Label key={index}>
+      <FiltersContainer itemsLength={colfilters.length}>
+        {colfilters.map((filter: TableFilter) => (
+          <Label key={filter.label}>
             {filter.multiselect ? (
               <>
                 {filter.label}
-                <StyledMultiSelect
-                  className="multiselect"
+                <SelectColumnFilter
+                  setFilter={setFilter}
                   options={filter.options}
-                  value={Array.isArray(filter.value) ? filter.value : []}
-                  onChange={filter.onChange}
-                  labelledBy="Select"
-                  hasSelectAll={false}
-                  overrideStrings={overrideStrings}
+                  setValue={filter.onChange}
+                  gotoPage={gotoPage}
                 />
               </>
             ) : (
               <>
-                {filter.label}
-                <Dropdown
-                  data-testid={filter.key}
-                  onChange={(event: React.ChangeEvent<HTMLInputElement>) =>
-                    filter?.onChange && filter.onChange(event.target.value)
-                  }
-                  value={typeof filter.value === 'string' ? filter.value : ''}
-                >
-                  {filter.options.map((option) => (
-                    <option
-                      value={option.value}
-                      key={Array.isArray(option.value) ? JSON.stringify(option.value) : option.value}
-                    >
-                      {option.label}
-                    </option>
-                  ))}
-                </Dropdown>
+                <div data-testid="select-col-filter">
+                  {filter.label}
+                  <Select
+                    className="basic-single"
+                    classNamePrefix="select"
+                    //@ts-ignore
+                    options={filter.options}
+                    onChange={(val: any) => filter.onChange && filter.onChange(val.value)}
+                    isSearchable={true}
+                    defaultValue={filter.options[0]}
+                    value={filter.options.find((op) => op.value === filter.value)}
+                  />
+                </div>
               </>
             )}
           </Label>
@@ -324,6 +375,11 @@ function Table({
     rightCol = searchCol;
   }
 
+  useEffect(() => {
+    onPage(pageIndex + 1);
+    setSelectedRow(undefined);
+  }, [pageIndex]);
+
   return (
     <>
       {showFilters && (
@@ -336,52 +392,92 @@ function Table({
           </Grid>
         </SectionHeader>
       )}
-
-      {showContent && (
-        <>
-          <StyledTable variant={variant}>
-            <ReactPlaceholder ready={!loading || false} showLoadingAnimation customPlaceholder={awesomePlaceholder}>
-              <thead>
-                <tr>
-                  {headers.map((header, index) => {
-                    return (
-                      <th key={index} style={header.style || {}}>
-                        {header.label}
+      <InfScroll loadMore={loadMoreItem} hasMore={hasMoreItem} loader={loader}>
+        <StyledTable variant={variant} {...getTableProps()} readOnly={readOnly}>
+          <thead>
+            {
+              // Loop over the header rows
+              headerGroups.map((headerGroup) => (
+                // Apply the header row props
+                <tr {...headerGroup.getHeaderGroupProps()} key={headerGroup.id}>
+                  {
+                    // Loop over the headers in each row
+                    headerGroup.headers.map((column) => (
+                      // Apply the header cell props
+                      <th {...column.getHeaderProps(column.getSortByToggleProps())} key={column.id}>
+                        {
+                          // Render the header
+                          column.render('Header')
+                        }
+                        &nbsp;
+                        <span>{getColumnSortedIcon(column.isSorted, column.isSortedDesc)}</span>
                       </th>
-                    );
-                  })}
+                    ))
+                  }
                 </tr>
-              </thead>
-              <tbody>{children}</tbody>
-            </ReactPlaceholder>
-          </StyledTable>
-          {pagination && rowCount > 0 && (
-            <Grid cols={12}>
-              <Grid.Row collapse="992" gutter={[]} align="center">
-                <Grid.Col span={8}>
-                  <StyledPagination>
-                    {onPage ? (
-                      <PaginationItemsDetail rowCount={rowCount} limit={limit} page={page} onPage={onPage} />
-                    ) : (
-                      <PaginationItems rowCount={rowCount} limit={limit} page={page} onPrev={onPrev} onNext={onNext} />
-                    )}
-                  </StyledPagination>
-                </Grid.Col>
-                {pageLimits?.length > 0 && (
-                  <Grid.Col span={4} style={{ textAlign: 'right' }}>
-                    <Dropdown
-                      style={{ display: 'inline-block', width: '160px' }}
-                      value={String(limit)}
-                      onChange={handlePageLimitChange}
+              ))
+            }
+          </thead>
+          <tbody {...getTableBodyProps()}>
+            <ReactPlaceholder ready={!loading || false} showLoadingAnimation customPlaceholder={awesomePlaceholder}>
+              {page.length > 0 ? (
+                page.map((row: any) => {
+                  prepareRow(row);
+                  // Use prop if passed from parent. Otherwise use component's internal state.
+                  const currentActiveSelector = activeSelector ?? selectedRow?.original[`${rowSelectorKey}`];
+                  return (
+                    <tr
+                      {...row.getRowProps()}
+                      className={currentActiveSelector === row?.original[`${rowSelectorKey}`] ? 'active' : ''}
+                      key={row?.id}
+                      onClick={() => updateSelectedRow(row)}
                     >
-                      {generateOptions(pageLimits)}
-                    </Dropdown>
-                  </Grid.Col>
-                )}
-              </Grid.Row>
-            </Grid>
-          )}
-        </>
+                      {row.cells.map((cell: Cell) => {
+                        return (
+                          <td {...cell.getCellProps()} key={cell.getCellProps().key}>
+                            {cell.render('Cell')}
+                          </td>
+                        );
+                      })}
+                    </tr>
+                  );
+                })
+              ) : (
+                <tr>
+                  <td colSpan={10}>{noDataFoundElement}</td>
+                </tr>
+              )}
+            </ReactPlaceholder>
+          </tbody>
+        </StyledTable>
+      </InfScroll>
+      {pagination && rowCount > 0 && (
+        <Grid cols={12}>
+          <Grid.Row collapse="992" gutter={[]} align="center">
+            <Grid.Col span={8}>
+              <StyledPagination>
+                <PaginationItemsDetail
+                  canNextPage={canNextPage}
+                  nextPage={nextPage}
+                  canPreviousPage={canPreviousPage}
+                  previousPage={previousPage}
+                  pageIndex={pageIndex}
+                  pageOptions={pageOptions}
+                />
+              </StyledPagination>
+            </Grid.Col>
+            <Grid.Col span={4}>
+              <div style={{ textAlign: 'right' }} data-testid="page-select">
+                <StyledSelect
+                  menuPosition="fixed"
+                  defaultValue={pageIndex || numOfItemsPerPage()[0]}
+                  options={numOfItemsPerPage()}
+                  onChange={(v: any) => handlePageLimitChange(v.value)}
+                ></StyledSelect>
+              </div>
+            </Grid.Col>
+          </Grid.Row>
+        </Grid>
       )}
     </>
   );
