+++ app/page-partials/my-dashboard/TeamInfoTabs/index.tsx
@@ -1,9 +1,8 @@
 import { useState, useEffect, useContext } from 'react';
 import { useRouter } from 'next/router';
 import styled from 'styled-components';
-import { Button as RequestButton, Tabs, Tab } from '@bcgov-sso/common-react-components';
-import Table from 'components/TableNew';
-import Button from 'html-components/Button';
+import { Tabs, Tab } from '@bcgov-sso/common-react-components';
+import Table from 'components/Table';
 import Dropdown from '@button-inc/bcgov-theme/Dropdown';
 import CenteredModal, { ButtonStyle } from 'components/CenteredModal';
 import TeamMembersForm, { Errors, validateTeam } from 'form-components/team-form/TeamMembersForm';
@@ -61,11 +60,6 @@ const TabWrapper = styled.div<{ marginTop?: string; marginBottom?: string; margi
 
 const RightFloat = styled.td`
   float: right;
-`;
-
-const PaddedButton = styled(Button)`
-  padding: 0 !important;
-  margin: 10px 0 !important;
 `;
 
 const CenteredTD = styled.td`
@@ -146,7 +140,7 @@ const ConfirmDeleteModal = ({
     case 'notAllowed':
       content = 'Before you delete the last team admin, you must assign a new admin.';
       props.confirmText = 'Okay';
-      props.buttonStyle = 'custom';
+      props.buttonStyle = 'primary';
   }
   return (
     <CenteredModal
@@ -438,11 +432,14 @@ function TeamInfoTabs({ alert, currentUser, team, loadTeams }: Props) {
         <Tab key="members" tab="Members">
           <TabWrapper>
             {isAdmin ? (
-              <PaddedButton>
-                <RequestButton onClick={() => setOpenAddTeamMemberModal(true)} data-testid="add-new-team-member">
-                  + Add New Team Members
-                </RequestButton>
-              </PaddedButton>
+              <button
+                className="primary"
+                style={{ margin: '1rem 0' }}
+                onClick={() => setOpenAddTeamMemberModal(true)}
+                data-testid="add-new-team-member"
+              >
+                + Add New Team Members
+              </button>
             ) : (
               <br />
             )}
@@ -641,7 +638,7 @@ function TeamInfoTabs({ alert, currentUser, team, loadTeams }: Props) {
                           getTeamServiceAccounts={fetchTeamServiceAccounts}
                         />
                       ) : (
-                        <RequestButton
+                        <button
                           style={{ marginBottom: 10 }}
                           onClick={async () => {
                             setLoading(true);
@@ -659,9 +656,10 @@ function TeamInfoTabs({ alert, currentUser, team, loadTeams }: Props) {
                               fetchTeamServiceAccounts(team.id);
                             }
                           }}
+                          className="primary"
                         >
                           + Request CSS API Account
-                        </RequestButton>
+                        </button>
                       )}
                     </Grid.Col>
                     <Grid.Col span={6}>
@@ -723,7 +721,7 @@ function TeamInfoTabs({ alert, currentUser, team, loadTeams }: Props) {
           />
         }
         onConfirm={onConfirmAdd}
-        buttonStyle="custom"
+        buttonStyle="primary"
         skipCloseOnConfirm={true}
         closable
       />
