+++ app/pages/admin-dashboard.tsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect, ReactNode } from 'react';
 import { useRouter } from 'next/router';
 import startCase from 'lodash.startcase';
 import { faTrash, faEdit, faEye, faTrashRestoreAlt } from '@fortawesome/free-solid-svg-icons';
-import Table from 'components/TableNew';
+import Table from 'components/Table';
 import { getRequestAll, deleteRequest, restoreRequest } from 'services/request';
 import { PageProps } from 'interfaces/props';
 import { Integration, Option } from 'interfaces/Request';
